cleaned up warnings

+++ include/abuf.h
@@ -4,7 +4,8 @@
 #define APPEND_BUF_INIT_SZ 1024
 #define APPEND_BUF_PAD 8
 #define APPEND_BUF_GROWTH_SZ 2
-
+#include <def.h>
+#include <unistd.h>
 struct append_buf {
   size_t sz;
   size_t cap;
@@ -19,6 +20,6 @@ void free_append_buffer(struct append_buf *buf);
 
 int write_buffer(struct append_buf *buf, char *new_data, size_t sz);
 
-int write_fd(struct append_buf *buf, int fd);
+ssize_t write_fd(struct append_buf *buf, int fd);
 
 #endif
